JWT Read, Budge Control, Memory Cache added

+++ trade.api/Models/DTOs/LoginDTOs/LoginGetDto.cs
@@ -4,6 +4,6 @@
     {
         public string UserName { get; set; }
         public string Token { get; set; }
-        public decimal Deposit { get; set; }
+        public double Deposit { get; set; }
     }
 }
